requirements.txt
git+https://github.com/DagOnStar/dagonstar
pytz
requests
python-magic
websockets
websocket-client
rel
libmagic
croniter
